Add files via upload
These files are packages for a node, that will create a topic to subscribe glove values from and then will use some code convert the gesture to a command, and send it to the drone topic that will control all or individual agents. Note: Looking at it now i should change the topic to be agent_topic instead of drone_topic. OR add another topic to refer to another or different agent.